requirements.txt
streamlit
python-dotenv
groq
scikit-learn
